+++ src/Classifiers/GaussianNB.php
@@ -182,6 +182,11 @@ class GaussianNB implements Estimator, Online, Probabilistic, Persistable
                 . ' Labeled training set.');
         }
 
+        if (in_array(DataFrame::CATEGORICAL, $dataset->types())) {
+            throw new InvalidArgumentException('This estimator only works with'
+                . ' continuous features.');
+        }
+
         $classes = $dataset->possibleOutcomes();
 
         $this->classes = $classes;