+++ tests/Classifiers/MultiLayerPerceptronTest.php
@@ -6,15 +6,16 @@ use Rubix\ML\Online;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Probabilistic;
-use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Datasets\Unlabeled;
 use Rubix\ML\NeuralNet\Layers\Dense;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
+use Rubix\ML\Datasets\Generators\Circle;
 use Rubix\ML\NeuralNet\Layers\Activation;
 use Rubix\ML\CrossValidation\Metrics\MCC;
 use Rubix\ML\Transformers\ZScaleStandardizer;
+use Rubix\ML\Datasets\Generators\Agglomerate;
 use Rubix\ML\Classifiers\MultiLayerPerceptron;
-use Rubix\ML\NeuralNet\ActivationFunctions\ELU;
+use Rubix\ML\NeuralNet\ActivationFunctions\ReLU;
 use Rubix\ML\NeuralNet\CostFunctions\CrossEntropy;
 use PHPUnit\Framework\TestCase;
 use InvalidArgumentException;
@@ -22,30 +23,28 @@ use RuntimeException;
 
 class MultiLayerPerceptronTest extends TestCase
 {
+    const TRAIN_SIZE = 400;
+    const TEST_SIZE = 5;
+    const MIN_PROB = 0.33;
+
+    protected $generator;
+
     protected $estimator;
-
-    protected $training;
-
-    protected $testing;
 
     public function setUp()
     {
-        $this->training = Labeled::load(dirname(__DIR__) . '/iris.dataset');
-
-        $transformer = new ZScaleStandardizer();
-
-        $transformer->fit($this->training);
-
-        $this->training->apply($transformer);
-
-        $this->testing = $this->training->randomize()->head(3);
+        $this->generator = new Agglomerate([
+            'inner' => new Circle(0., 0., 1., 0.01),
+            'middle' => new Circle(0., 0., 5., 0.05),
+            'outer' => new Circle(0., 0., 10., 0.1),
+        ]);
 
         $this->estimator = new MultiLayerPerceptron([
             new Dense(10),
-            new Activation(new ELU()),
-            new Dense(10),
-            new Activation(new ELU()),
-        ], 1, new Adam(0.001), 1e-4, new CrossEntropy(), 1e-3, new MCC(), 0.10, 3, 100);
+            new Activation(new ReLU()),
+            new Dense(5),
+            new Activation(new ReLU()),
+        ], 10, new Adam(0.01), 1e-4, new CrossEntropy(), 1e-3, new MCC(), 0.10, 3, 100);
     }
 
     public function test_build_classifier()
@@ -62,57 +61,43 @@ class MultiLayerPerceptronTest extends TestCase
         $this->assertEquals(Estimator::CLASSIFIER, $this->estimator->type());
     }
 
-    public function test_make_prediction()
+    public function test_train_partial_predict_proba()
     {
-        $this->estimator->train($this->training);
+        $dataset = $this->generator->generate(self::TRAIN_SIZE + self::TEST_SIZE);
 
-        $predictions = $this->estimator->predict($this->testing);
+        $transformer = new ZScaleStandardizer();
 
-        $this->assertEquals($this->testing->label(0), $predictions[0]);
-        $this->assertEquals($this->testing->label(1), $predictions[1]);
-        $this->assertEquals($this->testing->label(2), $predictions[2]);
+        $transformer->fit($dataset);
+        $dataset->apply($transformer);
 
-        $probabilities = $this->estimator->proba($this->testing);
+        $testing = $dataset->randomize()->take(self::TEST_SIZE);
 
-        $this->assertGreaterThanOrEqual(0.5, $probabilities[0][$this->testing->label(0)]);
-        $this->assertGreaterThanOrEqual(0.5, $probabilities[1][$this->testing->label(1)]);
-        $this->assertGreaterThanOrEqual(0.5, $probabilities[2][$this->testing->label(2)]);
-    }
-
-    public function test_partial_train()
-    {
-        $folds = $this->training->stratifiedFold(2);
+        $folds = $dataset->fold(3);
 
         $this->estimator->train($folds[0]);
-
         $this->estimator->partial($folds[1]);
+        $this->estimator->partial($folds[2]);
 
-        $predictions = $this->estimator->predict($this->testing);
+        foreach ($this->estimator->predict($testing) as $i => $prediction) {
+            $this->assertEquals($testing->label($i), $prediction);
+        }
 
-        $this->assertEquals($this->testing->label(0), $predictions[0]);
-        $this->assertEquals($this->testing->label(1), $predictions[1]);
-        $this->assertEquals($this->testing->label(2), $predictions[2]);
-
-        $probabilities = $this->estimator->proba($this->testing);
-
-        $this->assertGreaterThanOrEqual(0.5, $probabilities[0][$this->testing->label(0)]);
-        $this->assertGreaterThanOrEqual(0.5, $probabilities[1][$this->testing->label(1)]);
-        $this->assertGreaterThanOrEqual(0.5, $probabilities[2][$this->testing->label(2)]);
+        foreach ($this->estimator->proba($testing) as $i => $prob) {
+            $this->assertGreaterThan(self::MIN_PROB, $prob[$testing->label($i)]);
+        }
     }
 
     public function test_train_with_unlabeled()
     {
-        $dataset = new Unlabeled([['bad']]);
-
         $this->expectException(InvalidArgumentException::class);
 
-        $this->estimator->train($dataset);
+        $this->estimator->train(Unlabeled::quick());
     }
 
     public function test_predict_untrained()
     {
         $this->expectException(RuntimeException::class);
 
-        $this->estimator->predict($this->testing);
+        $this->estimator->predict(Unlabeled::quick());
     }
 }