+++ tests/PipelineTest.php
@@ -4,33 +4,38 @@ namespace Rubix\ML\Tests;
 
 use Rubix\ML\Online;
 use Rubix\ML\Pipeline;
+use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\MetaEstimator;
-use Rubix\ML\Datasets\Labeled;
-use Rubix\ML\Classifiers\GaussianNB;
-use Rubix\ML\Transformers\OneHotEncoder;
-use Rubix\ML\Transformers\ZScaleStandardizer;
-use Rubix\ML\Transformers\NumericStringConverter;
+use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Classifiers\NaiveBayes;
+use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\Datasets\Generators\Agglomerate;
+use Rubix\ML\Transformers\IntervalDiscretizer;
 use PHPUnit\Framework\TestCase;
+use InvalidArgumentException;
+use RuntimeException;
 
 class PipelineTest extends TestCase
 {
+    const TRAIN_SIZE = 200;
+    const TEST_SIZE = 5;
+    const MIN_PROB = 0.33;
+
+    protected $generator;
+
     protected $estimator;
-
-    protected $training;
-
-    protected $testing;
 
     public function setUp()
     {
-        $this->training = Labeled::load(__DIR__ . '/iris.dataset');
+        $this->generator = new Agglomerate([
+            'red' => new Blob([255, 0, 0], 3.),
+            'green' => new Blob([0, 128, 0], 1.),
+            'blue' => new Blob([0, 0, 255], 2.),
+        ]);
 
-        $this->testing = $this->training->randomize()->head(3);
-
-        $this->estimator = new Pipeline(new GaussianNB(), [
-            new NumericStringConverter(),
-            new OneHotEncoder(),
-            new ZScaleStandardizer(true),
+        $this->estimator = new Pipeline(new NaiveBayes(1.), [
+            new IntervalDiscretizer(6),
         ]);
     }
 
@@ -40,33 +45,41 @@ class PipelineTest extends TestCase
         $this->assertInstanceOf(MetaEstimator::class, $this->estimator);
         $this->assertInstanceOf(Online::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
+        $this->assertInstanceOf(Estimator::class, $this->estimator);
     }
 
-    public function test_make_prediction()
+    public function test_estimator_type()
     {
-        $this->estimator->train($this->training);
-
-        $predictions = $this->estimator->predict($this->testing);
-
-        $this->assertEquals($this->testing->label(0), $predictions[0]);
-        $this->assertEquals($this->testing->label(1), $predictions[1]);
-        $this->assertEquals($this->testing->label(2), $predictions[2]);
+        $this->assertEquals(Estimator::CLASSIFIER, $this->estimator->type());
     }
 
-    public function test_partial_train()
+    public function test_train_partial_predict_proba()
     {
-        $folds = $this->training->stratifiedFold(3);
+        $testing = $this->generator->generate(self::TEST_SIZE);
 
-        $this->estimator->train($folds[0]);
+        $this->estimator->train($this->generator->generate(self::TRAIN_SIZE / 2));
+        $this->estimator->partial($this->generator->generate(self::TRAIN_SIZE / 2));
 
-        $this->estimator->partial($folds[1]);
+        foreach ($this->estimator->predict($testing) as $i => $prediction) {
+            $this->assertEquals($testing->label($i), $prediction);
+        }
 
-        $this->estimator->partial($folds[2]);
+        foreach ($this->estimator->proba($testing) as $i => $prob) {
+            $this->assertGreaterThan(self::MIN_PROB, $prob[$testing->label($i)]);
+        }
+    }
 
-        $predictions = $this->estimator->predict($this->testing);
+    public function test_train_with_unlabeled()
+    {
+        $this->expectException(InvalidArgumentException::class);
 
-        $this->assertEquals($this->testing->label(0), $predictions[0]);
-        $this->assertEquals($this->testing->label(1), $predictions[1]);
-        $this->assertEquals($this->testing->label(2), $predictions[2]);
+        $this->estimator->train(Unlabeled::quick());
+    }
+
+    public function test_predict_untrained()
+    {
+        $this->expectException(RuntimeException::class);
+
+        $this->estimator->predict(Unlabeled::quick());
     }
 }